+++ concordium-consensus/src/Concordium/External.hs
@@ -361,6 +361,8 @@ startConsensus ::
     Word64 ->
     -- |Accounts table cache size
     Word32 ->
+    -- |Modules table cache size
+    Word32 ->
     -- |Serialized genesis data (c string + len)
     CString ->
     Int64 ->
@@ -399,6 +401,7 @@ startConsensus
     transactionsKeepAlive
     transactionsPurgingDelay
     accountsCacheSize
+    modulesCacheSize
     gdataC
     gdataLenC
     bidC
@@ -482,7 +485,8 @@ startConsensus
                   rpInsertionsBeforeTransactionPurge = fromIntegral insertionsBeforePurge,
                   rpTransactionsKeepAliveTime = TransactionTime transactionsKeepAlive,
                   rpTransactionsPurgingDelay = fromIntegral transactionsPurgingDelay,
-                  rpAccountsCacheSize = fromIntegral accountsCacheSize
+                  rpAccountsCacheSize = fromIntegral accountsCacheSize,
+                  rpModulesCacheSize = fromIntegral modulesCacheSize
                 }
 
 -- |Start up an instance of Skov without starting the baker thread.
@@ -500,6 +504,8 @@ startConsensusPassive ::
     -- |Number of seconds between transaction table purging runs
     Word64 ->
     -- |Accounts table cache size
+    Word32 ->
+    -- |Modules table cache size
     Word32 ->
     -- |Serialized genesis data (c string + len)
     CString ->
@@ -534,6 +540,7 @@ startConsensusPassive
     transactionsKeepAlive
     transactionsPurgingDelay
     accountsCacheSize
+    modulesCacheSize
     gdataC
     gdataLenC
     notifyContext
@@ -600,7 +607,8 @@ startConsensusPassive
                   rpInsertionsBeforeTransactionPurge = fromIntegral insertionsBeforePurge,
                   rpTransactionsKeepAliveTime = TransactionTime transactionsKeepAlive,
                   rpTransactionsPurgingDelay = fromIntegral transactionsPurgingDelay,
-                  rpAccountsCacheSize = fromIntegral accountsCacheSize
+                  rpAccountsCacheSize = fromIntegral accountsCacheSize,
+                  rpModulesCacheSize = fromIntegral modulesCacheSize
                 }
 
 -- |Shut down consensus, stopping any baker thread if necessary.
@@ -1318,6 +1326,8 @@ foreign export ccall
         Word64 ->
         -- |Accounts table cache size
         Word32 ->
+        -- |Modules table cache size
+        Word32 ->
         -- |Serialized genesis data (c string + len)
         CString ->
         Int64 ->
@@ -1360,6 +1370,8 @@ foreign export ccall
         -- |Number of seconds between transaction table purging runs
         Word64 ->
         -- |Accounts table cache size
+        Word32 ->
+        -- |Modules table cache size
         Word32 ->
         -- |Serialized genesis data (c string + len)
         CString ->