+++ concordium-consensus/src/Concordium/GlobalState/Basic/TreeState.hs
@@ -139,7 +139,7 @@ initialSkovData rp gd genState genTT mPending =
 -- This newtype establishes types for the @GlobalStateTypes@. The type variable @bs@ stands for the BlockState
 -- type used in the implementation.
 newtype PureTreeStateMonad bs m a = PureTreeStateMonad { runPureTreeStateMonad :: m a }
-  deriving (Functor, Applicative, Monad, MonadIO, BlockStateTypes, BS.AccountOperations,
+  deriving (Functor, Applicative, Monad, MonadIO, BlockStateTypes, BS.AccountOperations, BS.ModuleQuery,
             BS.BlockStateQuery, BS.BlockStateOperations, BS.BlockStateStorage, BS.ContractStateOperations, TimeMonad)
 
 deriving instance (MonadProtocolVersion m) => MonadProtocolVersion (PureTreeStateMonad bs m)