+++ concordium-consensus/src/Concordium/GlobalState/Persistent/BlockState/Updates.hs
@@ -58,7 +58,7 @@ migrateUpdateQueue f UpdateQueue{..} = do
     .. -- copy over the next sequence number
     }
 
-instance (MonadBlobStore m, Serialize e, MHashableTo m H.Hash e)
+instance (MonadBlobStore m, Serialize e)
         => BlobStorable m (UpdateQueue e) where
     storeUpdate uq@UpdateQueue{..} = do
         l <- forM uqQueue $ \(t, r) -> do
@@ -70,7 +70,6 @@ instance (MonadBlobStore m, Serialize e, MHashableTo m H.Hash e)
                 sequence_ $ fst <$> l
         let uq' = uq{uqQueue = snd <$> l}
         return (putUQ, uq')
-    store uq = fst <$> storeUpdate uq
     load = do
         uqNextSequenceNumber <- get
         l <- getLength
@@ -364,7 +363,6 @@ instance
                     >> putCooldownParametersQueue
                     >> putTimeParametersQueue
             return (putPU, newPU)
-    store pu = fst <$> storeUpdate pu
     load = do
         mRKQ <- label "Root keys update queue" load
         mL1KQ <- label "Level 1 keys update queue" load
@@ -571,7 +569,6 @@ instance (MonadBlobStore m, IsChainParametersVersion cpv)
                 pendingUpdates = pU
             }
         return (pKC >> pCPU >> pCP >> pPU, newUpdates)
-    store u = fst <$> storeUpdate u
     load = do
         mKC <- label "Current key collection" load
         mCPU <- label "Current protocol update" load