+++ concordium-consensus/src/Concordium/GlobalState/Types.hs
@@ -37,6 +37,8 @@ class BlockStateTypes (m :: Type -> Type) where
     -- |A reference type for 'BakerInfo'. This is used to avoid duplicating 'BakerInfo' in the
     -- state where possible.
     type BakerInfoRef m :: Type
+    -- |A reference to an instrumented module.
+    type InstrumentedModuleRef m :: WasmVersion -> Type
 
 -- |Data retrieved from the existing skov instance that is needed to construct
 -- the new instance. This is an existential type that closes over the **new**
@@ -77,6 +79,7 @@ instance BlockStateTypes (MGSTrans t m) where
     type Account (MGSTrans t m) = Account m
     type ContractState (MGSTrans t m) = ContractState m
     type BakerInfoRef (MGSTrans t m) = BakerInfoRef m
+    type InstrumentedModuleRef (MGSTrans t m) = InstrumentedModuleRef m
 
 deriving via MGSTrans MaybeT m instance BlockStateTypes (MaybeT m)
 deriving via MGSTrans (ExceptT e) m instance BlockStateTypes (ExceptT e m)