+++ concordium-consensus/tests/scheduler/SchedulerTests/RejectReasons.hs
@@ -93,7 +93,7 @@ transactionInputs = [
 
 type TestResult = ([(BlockItemWithStatus, Types.ValidResult)],
                    [(TransactionWithStatus, Types.FailureKind)],
-                   [(Types.ContractAddress, Instance)])
+                   [(Types.ContractAddress, Types.BasicInstance)])
 
 testRejectReasons :: IO TestResult
 testRejectReasons = do